Add GUI and spike implementation for #8 (Distribute rows/columns)
This is the current state of implementation:
It works, but only if you output anything to the transcript while reloading the invariant dimension scales, and a Transcript window is open. I have no clue why that's the case.
Another bug is that after distributing the cells, dragging any splitter will lead to unexpected behavior. This appears to be caused by a bug in #balanceOffsets implementation ProportionalSplitterMorph that is responsible for a small shift of every splitter whenever you mouse-down it. To test the current implementation, I temporarily apend the following line to ProportionalSplitterMorph>>#balanceOffsets:
	#iDoNotLikeBugs ifNotNil: [^ self].
Next steps are to find out why you need an open Transcript, contact Marcel Taeumel regarding to the splitter morphs and make the tests running.