Create alpha release for iteration §5
New features:
[*] Automatic cell saving (#17)
[*] Warning/error messages for bad sheet dimension inputs (#21)
[*] Button for deleting saved sheets (#27)

Add a prototype for simple formulas (#26)
The mechanism for differentiating between focused and non-focused cells is demonstrated by displaying a formula-like representation of a cell's content when it is not focused (obviously, for real formulas as described in #25 we would just display formulas when cells ARE focused).

Miscellaneous stability and functionality enhancements.